src/ObservabilityService/requirements.txt
Flask==3.0.2
requests==2.31.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
psutil
prometheus_client
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-requests